+++ src/components/ChipGroup.tsx
@@ -8,16 +8,14 @@ const styles = StyleSheet.create({
     flex: 1,
     height: 42,
     minHeight: 42,
-    paddingTop: 4,
-    paddingLeft: 16,
-    paddingBottom: 16,
-  } as ViewStyle,
-  ScrollViewWrapper: {
-    height: 36,
-    maxHeight: 36,
   } as ViewStyle,
   ScrollView: {
     flex: 1,
+    height: 42,
+    minHeight: 42,
+    paddingTop: 4,
+    paddingLeft: 16,
+    paddingBottom: 16,
   } as ViewStyle,
   ScrollViewItem: {
     marginRight: 8,
@@ -36,7 +34,6 @@ type ChipGroupProps = {
 
 type ChipGroupState = {
   chipStates: { [id: string]: boolean };
-  firstChipToggled: boolean;
 };
 
 export class ChipGroup extends React.PureComponent<ChipGroupProps, ChipGroupState> {
@@ -44,7 +41,6 @@ export class ChipGroup extends React.PureComponent<ChipGroupProps, ChipGroupStat
     super(props);
 
     this.state = {
-      firstChipToggled: false,
       chipStates: props.chips.reduce((prev, cp) => {
         prev[cp.id] = true;
         return prev;
@@ -56,12 +52,22 @@ export class ChipGroup extends React.PureComponent<ChipGroupProps, ChipGroupStat
 
   toggleActive(id: string, isActive: boolean) {
     const newChipStates = { ...this.state.chipStates };
-    if (!this.state.firstChipToggled) {
+    const chips = Object.values(newChipStates);
+    const active = Object.entries(newChipStates)
+      .filter(([_, v]) => v)
+      .map(([k]) => k);
+
+    /*
+     * The logic here is:
+     * - if all toggles are active, toggling one should disable all the other
+     * - if I'm disabling the last active toggle, all toggles should be enabled
+     */
+    if (active.length === chips.length || (active.length === 1 && active[0] === id)) {
       Object.keys(newChipStates).forEach((k) => {
-        newChipStates[k] = false;
+        newChipStates[k] = active.length < chips.length;
       });
       newChipStates[id] = true;
-      this.setState({ firstChipToggled: true, chipStates: newChipStates });
+      this.setState({ chipStates: newChipStates });
       this.props.onActiveChanged([id]);
     } else {
       newChipStates[id] = isActive;
@@ -77,20 +83,18 @@ export class ChipGroup extends React.PureComponent<ChipGroupProps, ChipGroupStat
   render() {
     return (
       <View style={styles.Wrapper}>
-        <View style={styles.ScrollViewWrapper}>
-          <ScrollView style={styles.ScrollView} horizontal>
-            {this.props.chips.map((chip) => (
-              <View key={`chip-${chip.id}`} style={styles.ScrollViewItem}>
-                <ChipToggle
-                  id={chip.id}
-                  chipProps={chip.props}
-                  active={this.state.chipStates[chip.id]}
-                  setActive={this.toggleActive}
-                />
-              </View>
-            ))}
-          </ScrollView>
-        </View>
+        <ScrollView style={styles.ScrollView} horizontal>
+          {this.props.chips.map((chip) => (
+            <View key={`chip-${chip.id}`} style={styles.ScrollViewItem}>
+              <ChipToggle
+                id={chip.id}
+                chipProps={chip.props}
+                active={this.state.chipStates[chip.id]}
+                setActive={this.toggleActive}
+              />
+            </View>
+          ))}
+        </ScrollView>
       </View>
     );
   }